Tidy up the main function and global namespace.
Instances of `return 1` in main were changed to the standard
`return EXIT_FAILURE`, and the single instance of `return 0`
was changed to `return EXIT_SUCCESS`.

The forward declaration of `ackermann` was also not needed, so it was
was removed and its default argument was placed on the single
declaration/definition of `ackermann`.

Author : Victor Scott Cushman
Email  : VictorSCushman@gmail.com